gcc whinning fixes
Former-commit-id: d5b050e603198c364ee22af3764f84e7539b9f15

+++ libfixwav/src/include/audio.h
@@ -1,7 +1,7 @@
 #ifndef _AUDIO_H
 #define _AUDIO_H
-
-
+#include "c_utils.h"
+#include "structures.h"
 #define AFMT_WAVE 1
 #define NO_AFMT_FOUND 4
 #define AFMT_WAVE_GOOD_HEADER 10
@@ -12,6 +12,6 @@
 int fixwav_getinfo(char* filename);
 
 int launch_sox(char** filename);
-int extract_audio_info(fileinfo_t *info, uint8_t * header);
+uint8_t extract_audio_info(fileinfo_t *info, uint8_t * header);
 
 #endif
